qmp/hmp: correct rocker flow packet counter
It appears that flow count statistics are recorded but not reported via
netlink as the former is done using the 'hits' field of the stats structure
of struct of_dpa_flow while the latter uses the rx_pkts field.

Resolve this problem by only using the rx_pkts field.
Also remove the hits field as it is no longer used.

Signed-off-by: Simon Horman <simon.horman@netronome.com>