qmp/hmp: add flow byte counters
A add tx and rx flow byte counters to complement the existing packet
counters. Report these new counters via netlink.

These counters are outside of the OF-DPA 1.0 and 2.0 specifications.

The motivation for this counter is to make it easier to offload Open
vSwitch (Linux kernel) datapath flows.  The Open vSwitch user-space (there
are others) assumes that flows have both packet and byte counters.

Like the tx packet counter the tx byte counter is reported via netlink
but is not incremented.

Signed-off-by: Simon Horman <horms@verge.net.au>