qmp/hmp: add flow idle counter
A add idle counter, which is currently calculated based on the current when
time a packet is received for a flow. Report this new counter via netlink.

This counter is outside of the OF-DPA 1.0 and 2.0 specifications.

The motivation for this counter is to make it easier to offload Open
vSwitch (Linux kernel) datapath flows whose life-cycle are managed from
user-space. The Open vSwitch user-space (there are others) use the their
most-recently primarily uses the used time to determine which flows to evict.

Signed-off-by: Simon Horman <horms@verge.net.au>